+++ Source/Widgets/CesiumInspector/CesiumInspector.js
@@ -117,28 +117,39 @@ define([
         var globeDepthFrustum = document.createElement('div');
         globeDepth.appendChild(globeDepthFrustum);
 
+        var pickDepth = document.createElement('div');
+        generalSection.appendChild(pickDepth);
+        var pCheckbox = document.createElement('input');
+        pCheckbox.type = 'checkbox';
+        pCheckbox.setAttribute('data-bind', 'checked: pickDepth, click: showPickDepth');
+        pickDepth.appendChild(pCheckbox);
+        pickDepth.appendChild(document.createTextNode('Show pick depth'));
+
+        var depthFrustum = document.createElement('div');
+        generalSection.appendChild(depthFrustum);
+
         // Use a span with HTML binding so that we can indent with non-breaking spaces.
         var gLabel = document.createElement('span');
         gLabel.setAttribute('data-bind', 'html: "&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Frustum:"');
-        globeDepthFrustum.appendChild(gLabel);
+        depthFrustum.appendChild(gLabel);
 
         var gText = document.createElement('span');
-        gText.setAttribute('data-bind', 'text: globeDepthFrustumText');
-        globeDepthFrustum.appendChild(gText);
+        gText.setAttribute('data-bind', 'text: depthFrustumText');
+        depthFrustum.appendChild(gText);
 
         var gMinusButton = document.createElement('input');
         gMinusButton.type = 'button';
         gMinusButton.value = '-';
         gMinusButton.className = 'cesium-cesiumInspector-pickButton';
-        gMinusButton.setAttribute('data-bind', 'click: decrementGlobeDepthFrustum');
-        globeDepthFrustum.appendChild(gMinusButton);
+        gMinusButton.setAttribute('data-bind', 'click: decrementDepthFrustum');
+        depthFrustum.appendChild(gMinusButton);
 
         var gPlusButton = document.createElement('input');
         gPlusButton.type = 'button';
         gPlusButton.value = '+';
         gPlusButton.className = 'cesium-cesiumInspector-pickButton';
-        gPlusButton.setAttribute('data-bind', 'click: incrementGlobeDepthFrustum');
-        globeDepthFrustum.appendChild(gPlusButton);
+        gPlusButton.setAttribute('data-bind', 'click: incrementDepthFrustum');
+        depthFrustum.appendChild(gPlusButton);
 
         // Primitives
         var prim = document.createElement('div');